CORE-494 - Make DatabaseInitializer able to cope with no tables to be dropped found.

+++ backend/src/main/java/quarano/core/DatabaseInitializer.java
@@ -3,6 +3,7 @@ package quarano.core;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
+import java.sql.DatabaseMetaData;
 import java.util.stream.Collectors;
 
 import javax.sql.DataSource;
@@ -17,6 +18,7 @@ import org.springframework.stereotype.Component;
 
 /**
  * @author Oliver Drotbohm
+ * @author Felix Schultze
  */
 @Profile("staging")
 @RequiredArgsConstructor
@@ -29,20 +31,26 @@ public class DatabaseInitializer implements FlywayConfigurationCustomizer {
 	 * @see org.springframework.boot.autoconfigure.flyway.FlywayConfigurationCustomizer#customize(org.flywaydb.core.api.configuration.FluentConfiguration)
 	 */
 	@Override
+	@SuppressWarnings("null")
 	public void customize(FluentConfiguration configuration) {
 
 		DataSource dataSource = configuration.getDataSource();
 
 		String url;
 		try {
-			url = JdbcUtils.extractDatabaseMetaData(dataSource, "getURL");
+
+			url = JdbcUtils.extractDatabaseMetaData(dataSource, DatabaseMetaData::getURL);
 
 			if (!url.contains("postgres")) {
 				log.info("Database is not a Postgres! Skipping data wiping!");
 				return;
 			}
 
-		} catch (MetaDataAccessException e) {}
+		} catch (MetaDataAccessException e) {
+
+			log.warn("Failed to lookup JDBC url for {}.", configuration);
+			return;
+		}
 
 		log.info("Wiping database tables.");
 
@@ -53,12 +61,20 @@ public class DatabaseInitializer implements FlywayConfigurationCustomizer {
 				.filter(it -> !it.startsWith("pg_") && !it.startsWith("sql_"))
 				.map(it -> String.format("\"%s\"", it))
 				.collect(Collectors.joining(", "));
+
+		if (droptables.isBlank()) {
+
+			log.info("No Quarano database tables found.");
+			return;
+		}
+
 		log.info("Dropping database tables " + droptables);
 		template.execute(String.format("DROP TABLE %s CASCADE;", droptables));
 
 		// https://github.com/yugabyte/yugabyte-db/issues/1822
 		log.info("Initializing flyway history table by our own");
-		var flywayTable = "CREATE TABLE \"public\".\"flyway_schema_history\" (" +
+
+		template.execute("CREATE TABLE \"public\".\"flyway_schema_history\" (" +
 				"\"installed_rank\" INT NOT NULL primary key," +
 				"\"version\" VARCHAR(50)," +
 				"\"description\" VARCHAR(200) NOT NULL," +
@@ -69,7 +85,6 @@ public class DatabaseInitializer implements FlywayConfigurationCustomizer {
 				"\"installed_on\" TIMESTAMP NOT NULL DEFAULT now()," +
 				"\"execution_time\" INTEGER NOT NULL," +
 				"\"success\" BOOLEAN NOT NULL" +
-				")";
-		template.execute(flywayTable);
+				")");
 	}
 }
